requirements.txt
asgiref==3.5.0
boto3==1.21.39
botocore==1.24.39
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cryptography==36.0.2
decorator==5.1.1
Django==4.0.4
django-rest-knox==4.2.0
djangorestframework==3.13.1
dynaconf==3.1.7
hvac==0.11.2
idna==3.3
Jinja2==3.1.1
jmespath==1.0.0
loguru==0.6.0
MarkupSafe==2.1.1
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
s3transfer==0.5.2
shortuuid==1.0.8
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
validators==0.18.2
win32-setctime==1.1.0
